+++ src/monomials/nibblemonom.jl
@@ -23,7 +23,7 @@ uppernibbles(a::NibbleMonom) = a.ev .& 0xf0
 monom_max_vars(a::NibbleMonom) = monom_max_vars(typeof(a))
 
 monom_max_vars(::Type{<:NibbleMonom{N}}) where N = 2*N
-monom_totaldeg(a::NibbleMonom) = a.deg % UInt
+monom_totaldeg(a::NibbleMonom) = a.deg % UInt8
 monom_copy(a::NibbleMonom) = a
 
 function monom_construct_hash_vector(rng::AbstractRNG, ::Type{NibbleMonom{N}}, ::Integer) where N
